+++ src/router/routes.js
@@ -6,7 +6,6 @@ const routes = [
     children: [
       { path: '', component: () => import('../pages/PageDashboard.vue') },
       { path: 'ibindex', component: () => import('../pages/PageIbindex.vue') },
-      { path: 'ibindex-weights', component: () => import('../pages/PageIbindexWeights.vue') },
       { path: 'funds', component: () => import('../pages/PageFunds.vue') },
     ]
   },